clean up for alt branch
