+++ groovy-webapp/src/app/app.module.ts
@@ -8,20 +8,20 @@ import { SignUpComponent } from './components/sign-up/sign-up.component';
 import { ProfileComponent } from './components/profile/profile.component';
 import { AboutComponent } from './components/about/about.component';
 import { HomeComponent } from './components/home/home.component';
-import { StoryComponent } from './components/story/story.component';
-import { StoryListComponent } from './components/story-list/story-list.component';
+import { StoryComponent } from './components/not-alt/story/story.component';
+import { StoryListComponent } from './components/not-alt/story-list/story-list.component';
 import {SharedModule} from "./shared/shared.module";
 import {FormsModule} from "@angular/forms";
 import {HttpClientModule,HTTP_INTERCEPTORS} from "@angular/common/http";
 import {UserTokenInterceptor} from "./services/user-token.interceptor";
-import { PlotListComponent } from './components/plot-list/plot-list.component';
-import { PlotComponent } from './components/plot/plot.component';
-import { PlotListViewComponent } from './components/plot-list-view/plot-list-view.component';
-import { StoryDashboardComponent } from './components/story-dashboard/story-dashboard.component';
-import { AltStoryListComponent } from './components/alt-story-list/alt-story-list.component';
-import { AltStoryComponent } from './components/alt-story/alt-story.component';
-import { AltPlotListComponent } from './components/alt-plot-list/alt-plot-list.component';
-import { AltPlotComponent } from './components/alt-plot/alt-plot.component';
+import { PlotListComponent } from './components/not-alt/plot-list/plot-list.component';
+import { PlotComponent } from './components/not-alt/plot/plot.component';
+import { PlotListViewComponent } from './components/not-alt/plot-list-view/plot-list-view.component';
+import { StoryDashboardComponent } from './components/not-alt/story-dashboard/story-dashboard.component';
+import { AltStoryListComponent } from './components/alt-stories/alt-story-list/alt-story-list.component';
+import { AltStoryComponent } from './components/alt-stories/alt-story/alt-story.component';
+import { AltPlotListComponent } from './components/alt-plots/alt-plot-list/alt-plot-list.component';
+import { AltPlotComponent } from './components/alt-plots/alt-plot/alt-plot.component';
 
 
 @NgModule({
