+++ groovy-webapp/src/app/components/alt-plots/alt-plot-list/alt-plot-list.component.html
