+++ groovy-webapp/src/app/components/alt-plots/alt-plot-list/alt-plot-list.component.ts
@@ -1,9 +1,9 @@
 import {Component, effect} from '@angular/core';
-import {IPlot, IStory} from "../../models/story.model";
+import {IPlot, IStory} from "../../../models/story.model";
 import {Observable, of} from "rxjs";
-import {PlotService} from "../../services/plot.service";
-import {StoryService} from "../../services/story.service";
-import {DnStoryService} from "../../services/dn-story.service";
+import {PlotService} from "../../../services/plot.service";
+import {StoryService} from "../../../services/story.service";
+import {DnStoryService} from "../../../services/dn-story.service";
 
 @Component({
   selector: 'app-alt-plot-list',
