+++ groovy-webapp/src/app/components/alt-plots/alt-plot/alt-plot.component.html
