+++ groovy-webapp/src/app/components/alt-stories/alt-story-list/alt-story-list.component.html
