+++ groovy-webapp/src/app/components/alt-stories/alt-story-list/alt-story-list.component.ts
@@ -1,16 +1,16 @@
-import {Component, computed, OnInit} from '@angular/core';
-import {defaultStory, IStory} from "../../models/story.model";
-//import {catchError, combineLatest, EMPTY, map, merge, Observable, Subject, tap} from "rxjs";
-import {StoryService} from "../../services/story.service";
+import {Component, computed, OnInit, signal, Signal} from '@angular/core';
+import {defaultStory, IStory} from "../../../models/story.model";
+import {toSignal} from "@angular/core/rxjs-interop";
+import {DnStoryService} from "../../../services/dn-story.service";
 
 @Component({
-  templateUrl: './story-list.component.html',
-  styleUrls: ['./story-list.component.css'],
-
+  selector: 'app-alt-story-list',
+  templateUrl: './alt-story-list.component.html',
+  styleUrls: ['./alt-story-list.component.css']
 })
-export class StoryListComponent implements OnInit{
+export class AltStoryListComponent implements OnInit {
 
-  activeStory = computed(() => this.storyService.currentStory);
+  activeStory = computed(() => this.storyService.currentDnStory);
   // @ts-ignore
   story: IStory = this.activeStory();
   stories: IStory[] | undefined;
@@ -18,33 +18,39 @@ export class StoryListComponent implements OnInit{
   errorMessages = ""
 
 
-  constructor(private storyService: StoryService) {
-   }
+  constructor(private storyService: DnStoryService) {
+  }
 
 
-ngOnInit(){
+  ngOnInit() {
     this.storyService.getStories().subscribe(data => {
-        this.stories = data;
-        if (!this.stories?.find(story => story.id == defaultStory.id)){
-          this.stories.push(defaultStory)
-        }
-      });
+      this.stories = data;
+      if (!this.stories?.find(story => story.id == defaultStory.id)) {
+        this.stories.push(defaultStory)
+      }
+    });
+
   }
 
   save(): void {
-
-    console.log("save pressed...saving,");
-    // this.storyService.updateStory(currentStory);
+    console.log("save pressed")
     this.storyService.updateStory();
   }
 
+
   select(id: string) {
+
     // @ts-ignore
-    console.log("Selected ID", id ? id : null);
+    console.log("Selected DN ID", id ? id : null);
     let selected = <IStory>this.stories?.find(story => story.id == id);
     this.storyService.updateCurrentStory(selected);
     // @ts-ignore
-    // this.storyService.setCurrentStory(this.story.id);
+    //console.log("Selected ID", story.id?id:null);
+    // @ts-ignore
+    // this.storyService.updateCurrentStory(story);
+    //this.storyService.setCurrentStory(this.story.id);
+    // this.story = this.stories?.find(story => story.id == this.storyService.currentDnStory().id);
+
   }
 
   delete(id: string): void {
@@ -56,7 +62,7 @@ ngOnInit(){
         this.stories?.splice(idx, 1);
         // @ts-ignore
         this.storyService.updateCurrentStory(this.stories[idx - 1] ? this.stories[idx - 1] : defaultStory);
-        this.story = this.storyService.currentStory();
+        this.story = this.storyService.currentDnStory();
       },
       (err: any) => console.log(err)
     );
@@ -78,3 +84,4 @@ ngOnInit(){
   }
 
 }
+
