+++ groovy-webapp/src/app/components/alt-stories/alt-story/alt-story.component.html
