+++ groovy-webapp/src/app/components/alt-stories/alt-story/alt-story.component.ts
@@ -1,7 +1,7 @@
 import {Component, effect} from '@angular/core';
-import {IStory} from "../../models/story.model";
-import {StoryService} from "../../services/story.service";
-import {DnStoryService} from "../../services/dn-story.service";
+import {IStory} from "../../../models/story.model";
+import {StoryService} from "../../../services/story.service";
+import {DnStoryService} from "../../../services/dn-story.service";
 
 @Component({
   selector: 'app-alt-story',
