+++ groovy-webapp/src/app/components/not-alt/plot-list-view/plot-list-view.component.html
