+++ groovy-webapp/src/app/components/not-alt/plot-list-view/plot-list-view.component.ts
@@ -1,5 +1,5 @@
 import {Component, EventEmitter, Input, Output} from '@angular/core';
-import {IPlot} from "../../models/story.model";
+import {IPlot} from "../../../models/story.model";
 
 @Component({
   selector: 'plot-list-view',
