+++ groovy-webapp/src/app/components/not-alt/plot-list/plot-list.component.html
