+++ groovy-webapp/src/app/components/not-alt/plot/plot.component.html
