+++ groovy-webapp/src/app/components/not-alt/plot/plot.component.ts
@@ -1,9 +1,9 @@
 import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
-import {IPlot} from "../../models/story.model";
+import {IPlot} from "../../../models/story.model";
 import {ActivatedRoute} from "@angular/router";
 import {parseMappings} from "@angular/compiler-cli/src/ngtsc/sourcemaps/src/source_file";
 import {async, Observable} from "rxjs";
-import {PlotService} from "../../services/plot.service";
+import {PlotService} from "../../../services/plot.service";
 
 @Component({
   selector: 'app-plot',
