+++ groovy-webapp/src/app/components/not-alt/story-dashboard/story-dashboard.component.html
