+++ groovy-webapp/src/app/components/not-alt/story-list/story-list.component.css
