+++ groovy-webapp/src/app/components/not-alt/story/story.component.html
